requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
sqlalchemy==2.0.29
cryptography==42.0.5
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
itsdangerous==2.1.2
slowapi==0.1.9
secure==0.3.0
aiofiles==23.2.1
python-dotenv==1.0.1
